dash/requirements.txt
dash
pandas
plotly
numpy
pillow
flask
redis
